// pos/src/app/models/m-scale-details.ts
export class MScaleDetails {
    constructor() { }
    id?: number;
    sizeID?: number;
    colorID?: number;
    scaleHeaderID?: number;
    scaleCode?: string;
    colorCode?: string;
    sizeCode?: string;
    description?: string;
    visualOrder?: number;
    sizeName?: string;
    sSID?: number;
    active?:boolean;
}